backend/requirements.txt
black==22.10.0
celery==5.2.7
Django==4.1.1
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework_simplejwt==5.2.1
drf_yasg==1.21.4
pylama==8.4.1
requests==2.28.1
